Crate spork

Spork

A cross-platform module for measuring CPU and memory usage for processes, threads, and children threads. This module currently supports Linux, Windows, and OS X.

Basic Usage

extern crate spork;

use spork::{
  Spork,
  Error,
  ErrorKind,
  Platform,
  StatType,
  Stats
};

let spork = match Spork::new() {
  Ok(s) => s,
  Err(e) => panic!("Error creating spork client! {:?}", e)
};

println!("Using platform {:?}", spork.platform());
println!("CPU cores: {:?}x @ {:?} Hz", spork.num_cores(), spork.clock_speed());
 
// get process stats 
let p_stats = match spork.stats(StatType::Process) {
  Ok(s) => s,
  Err(e) => panic!("Error polling process stats! {:?}", e)
};
println!("Process stats: {:?}", p_stats);

// get thread stats 
let t_stats = match spork.stats(StatType::Thread) {
  Ok(s) => s,
  Err(e) => panic!("Error polling thread stats! {:?}", e)
};
println!("Thread stats: {:?}", t_stats);

Structs

Spork
A struct to monitor CPU and memory usage.

Stats
A struct holding CPU and memory usage information.

Enums

Error
An error type describing the possible error cases for this module. If compiled with the feature compile_unimplemented certain functions will always return Unimplemented errors at runtime. See the feature notes for more information.

ErrorKind
The kind of error being created.

Platform
The current system's platform, such as Linux, Windows, etc.

StatType
An enum describing how to scope the CPU and memory data. Process reads CPU and memory usage across the entire process and can be used with stats_with_cpus, Children reads CPU and memory for child threads of the calling thread and can also be used with stats_with_cpus, and Thread reads CPU and memory for the calling thread only. On Linux or OS X (POSIX) see getrusage and clock_gettime for more information, and on Windows see GetProcessMemoryInfo and GetProcessTimes.