Enum spork::StatType

pub enum StatType {
    Process,
    Thread,
    Children,
}

An enum describing how to scope the CPU and memory data. Process reads CPU and memory usage across the entire process and can be used with stats_with_cpus, Children reads CPU and memory for child threads of the calling thread and can also be used with stats_with_cpus, and Thread reads CPU and memory for the calling thread only. On Linux or OS X (POSIX) see getrusage and clock_gettime for more information, and on Windows see GetProcessMemoryInfo and GetProcessTimes.

Variants
Process

Read usage across the entire process.

Thread

Read usage for the calling thread.

Children

Read usage across all child threads of the calling thread. This can mean different things on different platforms, and usually doesn't do what you want it to do. Use with caution, here be dragons.