Struct spork::Spork

pub struct Spork { /* fields omitted */ }

A struct to monitor CPU and memory usage.

Important Notes:

The stats and stats_with_cpus functions measure CPU usage by the time between calls (in ms), separated by the provided StatType. For example, a single thread calling stat with the StatType::Thread option in a loop every second will correctly measure usage across the previous 1 second interval for only that thread. However, if for example two threads running in loops with different delays both try to read stats with StatType::Process then the interval over which the CPU load is calculated will be the delta between whenever the function was last called in either thread. Make sure your program accounts for this by using the correct StatType such that multiple threads do not interfere with each other's intended sample rates. Using stat and stat_with_cpus with StatType::Thread across multiple concurrent threads will correctly measure the results for each thread individually.

TLDR: Be careful using stat and stat_with_cpus with StatType::Process across multiple threads running concurrently.

Methods

impl Spork

fn new() -> Result<Spork, Error>

Create a new Spork instance.

fn stats(&self, kind: StatType) -> Result<Stats, Error>

Get CPU and memory statistics in a Stats instance for the provided StatType assuming usage across only 1 CPU core.

let spork = Spork::new().unwrap();
let stats = spork.stats(StatType::Thread).unwrap();

println!("CPU: {}%, Memory: {} bytes, Cores: {}, Type: {}, Polled at: {}", 
  stats.cpu, stats.memory, stats.cores, stats.kind, stats.polled);

fn stats_with_cpus(&self,
kind: StatType,
cores: Option<usize>)
-> Result<Stats, Error>

Get CPU and memory statistics in a Stats instance for the provided StatType assuming usage across count CPU core(s). If None is provided then all available CPU cores will be considered, where applicable.

let spork = Spork::new().unwrap();
// read stats across all available CPU cores
let stats = spork.stats_with_cpus(StatType::Thread, None).unwrap();

println!("CPU: {}%, Memory: {} bytes, Cores: {}, Type: {}, Polled at: {}", 
  stats.cpu, stats.memory, stats.cores, stats.kind, stats.polled);

// read stats considering only 2 CPU cores
let stats = spork.stats_with_cpus(StatType::Thread, Some(2));

println!("CPU: {}%, Memory: {} bytes, Cores: {}, Type: {}, Polled at: {}", 
  stats.cpu, stats.memory, stats.cores, stats.kind, stats.polled);

fn platform(&self) -> Platform

Get the system type.

fn clock_speed(&self) -> u64

Get the CPU clock speed, in Hz.

fn num_cores(&self) -> usize

Get the number of CPU cores for your system.

fn drop_history(&self, kind: StatType) -> Option<Stats>

Clear the stats history for the process or calling thread. This library works by tracking the timestamp of the last stats poll, per thread, such that polls from different threads do not interfere with each other. However, the downside to this approach is that some extra data has to stick around. This function will delete the timestamp of the previous poll for the process or calling thread, and if the same thread decides to call stats or stats_with_cpus again it will use the process' uptime as the duration over which to calculate CPU usage for the next call. If a Spork instance is shared among multiple threads with short lifespans then it's a good idea to call this when those threads exit.

fn read_history(&self, kind: StatType) -> Option<Stats>

Read a copy of the most recently polled stats for kind.